fix(lint): roll back ten more unjustified workspace clippy allows; record the other ten as a measured residual [fgdb-gate-weakening-rollback-mthlh]
89b49d38 and 21ccbfd7 added 27 workspace-wide clippy allows with no evidence.
8ae77715 removed the six correctness/suspicious/perf ones. This removes the
ten style lints that fire at 3 sites or fewer, fixing all 18 sites or
allowing them locally with a reason:
- manual_next_back: 0 sites;
- unnecessary_cast (1): bulk_load's usize-to-usize cast dropped;
- redundant_pattern_matching (1): scalar_text asserts .is_ok();
- match_like_matches_macro (1): a fgdb-sim test match becomes matches!;
- len_zero (1): fgdb-sim VFS metadata.is_empty();
- single_char_add_str (2): fuzz generators insert a char;
- chunks_exact_to_as_chunks (3): the CLI key-file hex decode (a validated
  64-char line, no remainder) and two gql tests use as_chunks::<N>().0;
- doc_lazy_continuation (3): two test docs began with ">=5"/">=6", which
  markdown reads as a blockquote;
- manual_async_fn (3): three fgdb-repl test transports are now async fn
  impls of RPITIT trait methods;
- collapsible_match (3):
  - fgdb-order leadership's pure condition becomes a match guard;
  - fgdb-gql branch_text (a delimiter pop) and fgdb standing_query/grouped
    (work-meter charges) keep site-local allows, because a guard would hide
    those side effects.

The ten allows that remain are a MEASURED RESIDUAL, not a justification.
The table comment now records unique sites per lint (workspace --all-targets,
2026-09-25):
- result_large_err 263;
- unusual_byte_groupings 44;
- collapsible_if 38;
- type_complexity 34;
- err_expect 30;
- unnecessary_mut_passed 18;
- unnecessary_literal_unwrap 15;
- manual_is_multiple_of 14;
- manual_noop_waker 13;
- too_many_arguments 9;
- plus rustc's recursion_depth_exceeding_limit, 8.
The false "Correctness/suspicious/perf stay denied" claim is replaced.
fgdb-gate-weakening-rollback-mthlh stays open on the residual and on its
lint-table-law item.

This began as a 16-lint rollback. 8ae77715 landed the overlapping six while
it was under test. On replay every overlapping hunk kept 8ae77715's version,
the two sessions' fixes were equivalent (entry API, the same seed regrouping,
the same local allows), and this commit carries only the other ten.

Evidence (private worktree at origin/main 8ae77715 + this change):
- cargo clippy --workspace --all-targets --keep-going --locked -- -D warnings:
  exit 0.
- Every touched file was rustfmt-clean at HEAD and still is.
- The pre-replay 16-lint version passed cargo test --no-fail-fast for every
  touched crate (fgdb, fgdb-gql, fgdb-order, fgdb-repl, fgdb-sim,
  fgdb-chronicle, fgdb-cli): 743 targets, 4548 passed, 0 failed.
- On this exact replayed tree:
  - cargo test --no-fail-fast -p fgdb-order -p fgdb-sim -p fgdb-repl;
  - fgdb-gql lib + any_shortest_walk_text, fuzz_prepare_never_panics,
    graph_insert_text;
  - fgdb lib + bulk_load, scalar_text, gql_indexed_expansion,
    gql_property_index, fuzz_execute_never_panics;
  - fgdb-cli cli_robot.
  All exit 0: 173 targets, 2200 passed, 0 failed.

+++ crates/fgdb/tests/bulk_load.rs
@@ -326,7 +326,7 @@ async fn native_create(db: &mut Database, contexts: &PurposeContexts, rows: &[Bu
                     ),
                 )
                 .unwrap();
-            assert_eq!(stats.completed_statements as usize, chunk.len());
+            assert_eq!(stats.completed_statements, chunk.len());
             assert_eq!(
                 (stats.created_vertices + stats.created_edges) as usize,
                 chunk.len()
